fix: address all PR review comments
- Commit assets/js/life-money/{compute,chart,main}.js as source of truth
- Use ESM export + CommonJS compat block (matching repo convention)
- Fix division by zero in chart x-axis labels when yearsLeft=0 or points<=1
- Fix daysLeft=0 division in main.js (savingsPerDay guard)
- Fix finalMonthlyExpenses overshoot in playground (track last-used value)
- Add for attributes to all slider labels for accessibility
- Remove TODO comment from playground/index.html (modules now committed)
- Remove assets/js/life-money/ from .gitignore
- Simplify jest.setup.js: remove code generation (files are committed)
- Fix jest.global-setup.js header comment
- Remove duplicate pretest hook from package.json (Jest globalSetup handles it)
- Remove dead setup-dirs.js
- Remove BOM from babel.config.js
- Fix import paths in FILE_STRUCTURE.md (4 levels, not 5)
- Rewrite SETUP_SUMMARY.md: remove absolute Windows paths, fix import depth
