+++ .gitignore
@@ -12,8 +12,7 @@ vendor/
 .bundle/
 Gemfile.lock
 
-# Generated by jest.global-setup.js at test time
-assets/js/life-money/
+# Generated test artifacts
 test/unit/assets/js/life-money/
 
 # Test coverage
