+++ SETUP_SUMMARY.md
@@ -21,16 +21,16 @@
 
 ### ✓ Created Test Files
 
-**C:\dev\dekay2001.github.io\test\unit\assets\js\compute.test.js**
+**test/unit/assets/js/compute.test.js**
 - 15 unit tests for the `computeRunway()` function
 - Tests cover: years/days calculation, portfolio depletion logic, balance tracking
-- All tests reference: `../../../../../assets/js/life-money/compute.js`
+- Imports: `../../../../assets/js/life-money/compute.js`
 
-**C:\dev\dekay2001.github.io\test\unit\assets\js\chart.test.js**
+**test/unit/assets/js/chart.test.js**
 - 11 unit tests for the `drawChart()` function
 - Tests cover: canvas context setup, grid rendering, depletion markers
 - Uses jsdom environment for DOM testing
-- All tests reference: `../../../../../assets/js/life-money/chart.js`
+- Imports: `../../../../assets/js/life-money/chart.js`
 
 ### Source Modules (Auto-generated by jest.setup.js)
 
@@ -51,40 +51,24 @@ These will be created in `C:\dev\dekay2001.github.io\assets\js\life-money\`:
 
 ### How to Execute Tests
 
-**Option 1 (Recommended - via npm):**
+**Run via npm (from repo root):**
 ```bash
-cd C:\dev\dekay2001.github.io
-npm test -- --watchAll=false
-```
-
-**Option 2 (Direct Node.js execution):**
-```bash
-cd C:\dev\dekay2001.github.io
-node run-tests.js
-```
-
-**Option 3 (Windows batch file):**
-```cmd
-cd C:\dev\dekay2001.github.io
-run-tests.bat
+npm test
 ```
 
 ### Expected Test Results
 
 When executed, Jest should report:
 - **Passed:** All 26 tests (15 compute.test.js + 11 chart.test.js)
-- **Coverage:** Compute and Chart modules with 100% function coverage
 - **Test suites:** 2 passed
 
-The jest.setup.js file ensures all necessary directories and source files are created before any tests run, so the relative paths in the test files will resolve correctly.
+Source modules (`assets/js/life-money/`) are committed to the repository and imported directly by the test files.
 
 ### Verification Checklist
 
 ✓ Test files created with correct import paths
-✓ Jest configured with setup hook in package.json  
-✓ Source module code embedded in jest.setup.js
-✓ Standalone test runner script (run-tests.js) created
-✓ Batch file wrapper (run-tests.bat) created for Windows execution
-✓ All imports use relative paths from test directory to assets/js/life-money
+✓ Jest configured in package.json
+✓ Source modules committed under `assets/js/life-money/`
+✓ All imports use relative paths: `../../../../assets/js/life-money/`
 
 The setup is complete and ready for test execution!
